+++ rings/src/main/java/cc/redberry/rings/Rings.java
@@ -100,6 +100,26 @@ public final class Rings {
     }
 
     /**
+     * Algebraic number field generated by the specified minimal polynomial
+     */
+    public static <Poly extends IUnivariatePolynomial<Poly>> AlgebraicNumberField<Poly>
+    AlgebraicExtension(Poly minimalPoly) {
+        return new AlgebraicNumberField<>(minimalPoly);
+    }
+
+    /**
+     * Quotient ring {@code baseRing/<modulus> }
+     *
+     * @deprecated Use either {@link #GF(IUnivariatePolynomial)} or {@link #AlgebraicExtension(IUnivariatePolynomial)}
+     */
+    @Deprecated
+    public static <uPoly extends IUnivariatePolynomial<uPoly>>
+    UnivariateQuotientRing<uPoly>
+    UnivariateQuotientRing(uPoly modulus) {
+        return modulus.isOverFiniteField() ? GF(modulus) : AlgebraicExtension(modulus);
+    }
+
+    /**
      * Ring of univariate polynomials over specified coefficient ring
      *
      * @param coefficientRing the coefficient ring
@@ -296,20 +316,11 @@ public final class Rings {
     }
 
     /**
-     * Quotient ring {@code baseRing/<modulus> }
-     */
-    public static <uPoly extends IUnivariatePolynomial<uPoly>>
-    UnivariateQuotientRing<uPoly>
-    UnivariateQuotientRing(IPolynomialRing<uPoly> baseRing, uPoly modulus) {
-        return new UnivariateQuotientRing<>(baseRing, modulus);
-    }
-
-    /**
      * Quotient ring {@code baseRing/<ideal> }
      */
     public static <Term extends AMonomial<Term>, Poly extends AMultivariatePolynomial<Term, Poly>>
     QuotientRing<Term, Poly>
-    QuotientRing(IPolynomialRing<Poly> baseRing, Ideal<Term, Poly> ideal) {
+    QuotientRing(MultivariateRing<Poly> baseRing, Ideal<Term, Poly> ideal) {
         return new QuotientRing<>(baseRing, ideal);
     }
 }
